+++ dao/src/main/java/dao/interfaces/GenericDao.java
@@ -5,16 +5,15 @@ import java.util.List;
 /*
 Generic dao interface. Include CRUD methods for all entities.
  */
-
 public interface GenericDao<T> {
 
     void create(T entity);
 
-    T read(int id);
+    T read(long id);
 
     void update(T entity);
 
-    void delete(int id);
+    void delete(long id);
 
     List<T> getAll();
 }
